Refactor test paths to remove unnecessary directory references
